Zipped AnnData object
Added a zipped AnnData object for anyone interested in playing around with the code. The AnnData object is with same cells as the pilot set "Pilot_2_rule_them_ALL.h5ad" but with about half the amount of genes for size reduction. For information on how this object was created see "Data_4_Github.Rmd" in the code folder